+++ slack-api-model/src/test/java/test_locally/util/ExecutorServiceFactoryTest.java
@@ -1,7 +1,6 @@
 package test_locally.util;
 
 import com.slack.api.util.thread.ExecutorServiceFactory;
-import com.slack.api.util.thread.ExecutorServiceProvider;
 import org.junit.Test;
 
 import java.util.concurrent.ExecutorService;
@@ -12,22 +11,6 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import static org.junit.Assert.assertTrue;
 
 public class ExecutorServiceFactoryTest {
-
-    @Test
-    public void getInstanceId() throws Exception {
-        ExecutorServiceProvider provider = new ExecutorServiceProvider() {
-            @Override
-            public ExecutorService createThreadPoolExecutor(String threadGroupName, int poolSize) {
-                return null;
-            }
-
-            @Override
-            public ScheduledExecutorService createThreadScheduledExecutor(String threadGroupName) {
-                return null;
-            }
-        };
-        assertTrue(provider.getInstanceId().matches("test_locally.util.ExecutorServiceFactoryTest\\$1\\@\\w+$"));
-    }
 
     @Test
     public void createDaemonThreadScheduledExecutor() throws InterruptedException {
